Update Auth Backend

+++ libs/backend/expedition/src/lib/expedition/expedition.controller.ts
@@ -6,6 +6,7 @@ import {
     Param,
     Post,
     Put,
+    Req,
     UseGuards
 } from '@nestjs/common';
 import { ExpeditionService } from './expediton.service';
@@ -15,7 +16,8 @@ import {
     ContinentEnum,
     IExpedition,
     IActivity,
-    IRole
+    IRole,
+    IUser
 } from '@avans-nx-expedition/shared/api';
 import {
     CreateExpeditionDto,
@@ -23,6 +25,8 @@ import {
     UpsertExpeditionDto
 } from '@avans-nx-expedition/backend/dto';
 import { Expedition } from './expedition.schema';
+import { Request } from 'express';
+import { AuthGuard } from '@avans-nx-expedition/backend/auth';
 
 @Controller('expedition')
 export class ExpeditionController {
@@ -53,34 +57,77 @@ export class ExpeditionController {
         return this.expeditionService.findOne(id);
     }
 
+    @UseGuards(AuthGuard)
     @Delete(':id')
-    async delete(@Param('id') id: string): Promise<IExpedition | null> {
+    async delete(
+        @Param('id') id: string,
+        @Req() req: Request
+    ): Promise<IExpedition | null> {
+        const user = (req as any)['user'];
+        const userId = user.user_id;
+
+        const expedition = await this.expeditionService.findOne(id);
+        if (!expedition) {
+            console.log('Expedition not found');
+            return Promise.resolve(null);
+        }
+
+        const expeditionOrganizerId = (
+            expedition.organizer as IUser
+        )._id.toString();
+        if (expeditionOrganizerId !== userId) {
+            console.log(
+                `OrganiserId ${expeditionOrganizerId} !== userId ${userId}`
+            );
+            console.log('User is not the organizer of this expedition');
+            return null;
+        }
+
         return this.expeditionService.delete(id);
     }
 
     @Post('')
-    // @UseGuards(expeditionExistGuard) NOT IMPLEMENTED YET
     create(@Body() expedition: CreateExpeditionDto): Promise<IExpedition> {
         return this.expeditionService.create(expedition);
     }
 
     @Post('activity')
-    // @UseGuards(expeditionExistGuard) NOT IMPLEMENTED YET
     createActivity(@Body() activity: IActivity): Promise<IActivity> {
         return this.expeditionService.createActivity(activity);
     }
 
     @Post('role')
-    // @UseGuards(expeditionExistGuard) NOT IMPLEMENTED YET
     createRole(@Body() role: IRole): Promise<IRole> {
         return this.expeditionService.createRole(role);
     }
 
+    @UseGuards(AuthGuard)
     @Put(':id')
-    update(
+    async update(
         @Param('id') id: string,
-        @Body() expedition: UpdateExpeditionDto
+        @Body() expedition: UpdateExpeditionDto,
+        @Req() req: Request
     ): Promise<IExpedition | null> {
+        const user = (req as any)['user'];
+        const userId = user.user_id;
+
+        const expeditionDB = await this.expeditionService.findOne(id);
+        if (!expeditionDB) {
+            console.log('Expedition not found');
+            return Promise.resolve(null);
+        }
+
+        const expeditionOrganizerId = (
+            expeditionDB.organizer as IUser
+        )._id.toString();
+        if (expeditionOrganizerId !== userId) {
+            console.log(
+                `OrganiserId ${expeditionOrganizerId} !== userId ${userId}`
+            );
+            console.log('User is not the organizer of this expedition');
+            return Promise.resolve(null);
+        }
+
         return this.expeditionService.update(id, expedition);
     }
